Pages: Expose project upload logic to the unstable_pages API
