+++ .changeset/cool-toys-divide.md
@@ -1,0 +1,5 @@
+---
+"wrangler": minor
+---
+
+Pages: Expose project upload logic to the unstable_pages API
