+++ packages/wrangler/src/api/pages/index.ts
@@ -1,5 +1,7 @@
 import { deploy } from "./deploy";
+import { uploadDeploymentFiles } from "./uploadDeploymentFiles";
 
 export const unstable_pages = {
 	deploy,
+	uploadDeploymentFiles,
 };
